pyproject.toml
[tool.poetry]
name = "hw11"
version = "0.1.0"
description = ""
authors = ["taras55001 <taras5500@gmail.com>"]
license = "mit"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.100.0"}
uvicorn = {extras = ["standard"], version = "^0.23.1"}
sqlalchemy = "^2.0.19"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
